:recycle: Rename TokenRow component and styles to a more representative name
